feat: Add dropdown_button2 dependency
This commit adds the dropdown_button2 dependency with version 2.3.9. It is a transitive dependency and was retrieved from the hosted source on pub.dev.

The pubspec.yaml file has also been updated with the flutter_form_builder dependency version 9.1.1.

Additionally, the assets/fonts section in the pubspec.yaml file has been expanded to include the Poppins font family with various weights and styles.