Refactor PrimaryButton with icon support
This commit introduces changes to the PrimaryButton class to add support for displaying an icon. The button can now have an optional icon, which can be positioned on the left or right side of the button's text. The iconPosition property determines the position of the icon, and hasIcon property is used to specify whether the button has an icon or not. The commit also includes improvements to the code organization and readability.