Create SchoolAppPrimaryButton with specific properties
This commit creates a button called SchoolAppPrimaryButton with the text "Cliquez ici". It has a blue background, solid decoration, height of 50 pixels, an icon positioned to the right of the text, and white colored icon or text. Additionally, it includes an onTap function that will be executed when the button is clicked.