Delete app_chip.dart
This commit deletes the file app_chip.dart, which contained a class named SchoolAppChip. This class was responsible for displaying a chip widget with various customization options like label, width, height, icon, background color, and chip type. The chip type can be none, icon, or dot, each having its own rendering logic. The chip was built using a container with a specified width and height, decorated with a shape and background color. Inside the container, there was a row of button-like children widgets, including an icon or dot and a label.