feat: Add CustomCard and CustomPopupMenu widgets
This commit adds a new file `custom_card.dart` which contains the `CustomCard` widget. It also imports necessary packages and includes the widget implementation.

Additionally, the file `custom_pop_up.dart` is removed as it is no longer needed.

The `CustomPopupMenu` widget from the removed file is now replaced with `SchoolapPopMenuButton` widget from the `schoolap_pop_menu_button` file.

These changes improve the code structure and enhance the modularity of the project.